+++ .env
@@ -1,0 +1,1 @@
+HTTP_PORT=3000
